+++ src/MediaCollections/Models/Collections/MediaCollection.php
@@ -49,4 +49,24 @@ class MediaCollection extends Collection implements Htmlable
             ];
         })->keyBy('uuid')));
     }
+
+    public function jsonSerialize()
+    {
+        if (!($this->formFieldName ?? $this->collectionName)) {
+            return [];
+        }
+
+        return old($this->formFieldName ?? $this->collectionName) ?? $this->map(function (Media $media) {
+            return [
+                'name' => $media->name,
+                'file_name' => $media->file_name,
+                'uuid' => $media->uuid,
+                'preview_url' => $media->hasGeneratedConversion('preview') ? $media->getUrl('preview') : '',
+                'order' => $media->order_column,
+                'custom_properties' => $media->custom_properties,
+                'extension' => $media->extension,
+                'size' => $media->size,
+            ];
+        })->keyBy('uuid');
+    }
 }