+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to `laravel-medialibrary` will be documented in this file
 ## 9.0.0 - Unreleased
 
 - names of the generated conversions will now be put in a dedicated `generated_conversions` on media
-- adds support for media library pro
+- add support for media library pro
+- responsive image files can now be named using the `file_namer` key in the `media-library` config file (#2114)
 
 ## 8.10.1 - 2020-10-05